# internal/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Dockmesh API
  summary: Container management platform — HTTP API
  description: |
    REST API for Dockmesh. All endpoints are versioned under `/api/v1`.

    ## Authentication

    Two bearer-token flavours are accepted on the `Authorization` header:

    - **User JWTs** — obtained via `POST /auth/login`, short-lived (15 min),
      refreshable via `POST /auth/refresh`. Use for interactive UI / CLI sessions.
    - **API tokens** — prefix `dmt_`, long-lived, created under
      Settings → API Tokens. Use for CI, scripts, external integrations.

    Both share the same permission model — a token's role dictates what it can do.

    ## Versioning

    Version `v1` is stable in the sense that we don't remove fields or change
    response shapes without warning. New fields may be added at any time, so
    clients should ignore unknown fields gracefully.

    ## Errors

    Non-2xx responses use a uniform shape: `{"error": "<human-readable message>"}`.
    HTTP status carries the category (400/401/403/404/409/500).

    ## Maintenance

    This spec is hand-maintained. `TestOpenAPIDriftAgainstRoutes` compares the
    registered chi routes against the operations declared here and fails CI on
    mismatch — so every PR that touches a route must also touch this file.
  version: "1.0.0"
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
  contact:
    name: Dockmesh
    url: https://dockmesh.dev
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /agents/{id}/upgrade:
    post:
      operationId: agents.upgrade
      tags: [agents]
      summary: Trigger a self-upgrade on a connected agent
      description: |
        The agent downloads a fresh `dockmesh-agent-linux-<arch>` binary
        from the server's `/install/` path and restarts its systemd unit.
        Agent must be currently connected.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Upgrade dispatched.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: upgrading }
                  version: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503":
          description: Agent is not currently connected.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /containers:
    get:
      operationId: containers.list
      tags: [containers]
      summary: List containers across one or all hosts
      description: |
        Fan-out capable via `?host=all`. Each row is a Docker SDK
        `Container` summary plus `host_id` / `host_name` when fanning out.
      parameters:
        - $ref: "#/components/parameters/HostQuery"
        - name: all
          in: query
          schema: { type: boolean, default: false }
          description: Include stopped containers.
      responses:
        "200":
          description: Container list. Shape differs between single-host (array) and fan-out (FanOutResponse).
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items: { $ref: "#/components/schemas/Container" }
                  - $ref: "#/components/schemas/FanOutResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503":
          description: Docker daemon is unreachable on the target host.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /containers/{id}:
    parameters:
      - $ref: "#/components/parameters/ContainerIDPath"
      - $ref: "#/components/parameters/HostQuery"
    get:
      operationId: containers.inspect
      tags: [containers]
      summary: Inspect one container
      responses:
        "200":
          description: Full ContainerJSON from the Docker daemon.
          content:
            application/json:
              schema: { type: object, description: Passthrough of the Docker SDK ContainerJSON. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: containers.remove
      tags: [containers]
      summary: Remove a container
      parameters:
        - name: force
          in: query
          schema: { type: boolean, default: false }
          description: Kill running containers before removing.
      responses:
        "204": { description: Removed. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /volumes/{name}/browse/file:
    get:
      operationId: volumes.browse_file
      tags: [volumes]
      summary: Read a single file inside a volume, capped at 1 MiB (P.11.8)
      description: |
        Returns up to 1 MiB of content. `truncated: true` signals the file
        is larger — the UI then offers a download link instead of an inline
        preview. `binary: true` is set when a NUL byte is found in the
        first 8 KiB (same heuristic git uses).
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/HostQuery"
        - name: path
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: File content.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VolumeFileResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # -----------------------------------------------------------------------
  # networks — CRUD + topology view + prune
  # -----------------------------------------------------------------------
...
  /invite/{token}:
    parameters:
      - in: path
        name: token
        required: true
        schema: { type: string }
    get:
      operationId: invites.preview
      tags: [auth]
      summary: Preview an invite (public — used by the accept page)
      security: []
      responses:
        "200":
          description: Invite preview.
          content:
            application/json:
              schema:
                type: object
                properties:
                  role: { type: string }
                  scope_tags: { type: array, items: { type: string } }
                  email_hint: { type: string }
                  expires_at: { type: string, format: date-time }
        "404": { description: invite not found }
        "410": { description: expired or already used }
...
  /containers/{id}/files:
    get:
      operationId: containers.files.browse
      tags: [containers]
      summary: List one directory level inside a container's filesystem
      description: |
        Tar-streams the requested path out of the container and walks
        the archive at depth 1. Path must be absolute. Audited on every
        call because reading container files is a sensitive operation.
      parameters:
        - $ref: "#/components/parameters/ContainerIDPath"
        - $ref: "#/components/parameters/HostQuery"
        - name: path
          in: query
          schema: { type: string, default: "/" }
          description: Absolute path inside the container.
      responses:
        "200":
          description: Directory entries.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/VolumeEntry" }
        "400":
          description: Path is empty, not absolute, or points to a non-directory.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      operationId: containers.files.write
      tags: [containers]
      summary: Upload a single file into the container at an absolute path
      description: |
        Requires `containers.exec` because writing a file into a running
        container is effectively code injection. Parent directory must
        already exist. Content is base64-encoded so binary uploads work
        over plain JSON.
      parameters:
        - $ref: "#/components/parameters/ContainerIDPath"
        - $ref: "#/components/parameters/HostQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path, content]
              properties:
                path:
                  type: string
                  description: Absolute path inside the container.
                content:
                  type: string
                  description: Base64-encoded file content.
                mode:
                  type: integer
                  description: Unix permission bits; defaults to 0644 (420).
      responses:
        "200":
          description: File written.
          content:
            application/json:
              schema:
                type: object
                properties:
                  path: { type: string }
                  bytes: { type: integer }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /containers/{id}/files/download:
    get:
      operationId: containers.files.download
      tags: [containers]
      summary: Stream a file's raw bytes out of the container for download
      description: |
        Streams `application/octet-stream` with `Content-Disposition:
        attachment` so the browser triggers a file save. Used for binary
        files and files larger than the preview cap.
      parameters:
        - $ref: "#/components/parameters/ContainerIDPath"
        - $ref: "#/components/parameters/HostQuery"
        - name: path
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Raw file bytes.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # -----------------------------------------------------------------------
  # SSO additions (Tier C) — SAML, LDAP, OAuth2 alongside existing OIDC.
  # Plus auth-policy + avatar + alerts/global-env enhancements from Tier A/B.
  # -----------------------------------------------------------------------
...
  /hosts/{id}/stats/containers:
    get:
      operationId: hosts.batch_container_stats
      tags: [containers]
      summary: Latest CPU + memory sample per container (used by list view)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Array of latest samples.
          content:
            application/json:
              schema:
                type: array
                items: { type: object, additionalProperties: true }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "JWT or dmt_<40-char-base64url>"
      description: |
        Accepts either a user session JWT (from `POST /auth/login`) or an API
        token (prefix `dmt_`, from Settings → API Tokens). The route layer
        checks role permissions after authentication.

  # Schemas used across multiple groups live here. Group-specific schemas
  # live inline under the operation to keep the spec readable when new
  # groups get added.
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: invalid credentials
...
    VolumeEntry:
      type: object
      description: One row in a volume directory listing (P.11.8).
      required: [name, type, size, mode, mod_time]
      properties:
        name: { type: string }
        type:
          type: string
          enum: [file, dir, symlink]
        size: { type: integer, format: int64 }
        mode: { type: string, description: Unix-style "rwxr-xr-x" string. }
        mod_time: { type: string, format: date-time }
        link_dest: { type: string, description: For symlinks, the target path. }
    VolumeFileResult:
      type: object
      description: Inline file read from a volume (P.11.8).
      required: [content, size, truncated, binary]
      properties:
        content:
          type: string
          format: byte
          description: Base64-encoded file bytes — capped at 1 MiB.
        size:
          type: integer
          format: int64
          description: Actual file size on disk (may exceed content length when truncated).
        truncated: { type: boolean }
        binary:
          type: boolean
          description: True when the first 8 KiB contain a NUL byte.
...
    Container:
      type: object
      description: |
        Docker container summary. Mirrors the Docker SDK's
        `types.Container` with `host_id` + `host_name` added when the
        response comes from a fan-out list. Most fields are passthrough.
      properties:
        Id: { type: string }
        Names:
          type: array
          items: { type: string }
        Image: { type: string }
        ImageID: { type: string }
        Command: { type: string }
        Created: { type: integer, format: int64 }
        State:
          type: string
          enum: [created, restarting, running, removing, paused, exited, dead]
        Status: { type: string, example: "Up 5 minutes" }
        Labels:
          type: object
          additionalProperties: { type: string }
        host_id: { type: string }
        host_name: { type: string }
    FanOutResponse:
      type: object
      description: |
        Returned by list endpoints when `?host=all`. `items` is the flat
        aggregated list; `errors` lists hosts that failed to respond so
        the UI can show partial results with a warning banner.
      required: [items]
      properties:
        items:
          type: array
          items: { type: object }
          description: Rows from every host that answered. Shape depends on the endpoint.
        errors:
          type: array
          items:
            type: object
            properties:
              host_id: { type: string }
              host_name: { type: string }
              error: { type: string }
...
  parameters:
    HostQuery:
      name: host
      in: query
      description: "Target host id. Empty or `local` = central server's daemon; otherwise an agent UUID. `all` fans out to every online host."
      required: false
      schema:
        type: string
        default: local
...
    ContainerIDPath:
      name: id
      in: path
      description: Container id (long or short form) or name.
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but the token's role lacks the required permission.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Malformed request body or invalid parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...